+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v0.1.1 (2023-11-16)
 
+- Put sub_account_id inside the headers parameters
 - Develop `ExXendit.XenPlatform.create_account`
+- Develop `ExXendit.EWallet.create_charge`
 
 ## v0.1.0 (2023-11-16)